add selectors for each components
