+++ src/widgets/modules/webphone/presentation/CallInfo/selector.js
@@ -1,0 +1,28 @@
+import LPN from 'google-libphonenumber';
+
+const phoneUtil = LPN.PhoneNumberUtil.getInstance();
+
+function clean(str) {
+  return str.slice(0, str.indexOf('@'));
+}
+
+function getNationalPhone(raw, country = 'US') {
+  if (!raw) return '';
+  return phoneUtil.format(
+    phoneUtil.parse(
+      raw,
+      country
+    ),
+    LPN.PhoneNumberFormat.NATIONAL
+  );
+}
+
+export default (state, props, phone) => (
+  {
+    callingNumber: getNationalPhone(
+                  state.common.webphone.remoteIdentity ?
+                  clean(state.common.webphone.remoteIdentity.friendlyName) :
+                  state.common.webphone.toNumber
+                ),
+  }
+);
