+++ src/widgets/modules/webphone/presentation/Transfer/selector.js
@@ -1,0 +1,5 @@
+export default (state, props, phone) => (
+  {
+    transfer: (...args) => phone.webphone.transfer(...args),
+  }
+);
